test(refactoring): #2530 Fix test utils refs.
