+++ projects/igniteui-angular/src/lib/services/exporter-common/components-declarations.spec.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { IgxGridComponent } from '../../grid/grid.component';
+import { IgxGridComponent } from '../../grids/grid/grid.component';
 import { ExportTestDataService } from '../excel/test-data.service.spec';
 
 @Component({
