+++ projects/igniteui-angular/src/lib/services/exporter-common/test-methods.spec.ts
@@ -1,6 +1,6 @@
 
 import { TestBed } from '@angular/core/testing';
-import { IgxGridComponent } from '../../grid/grid.component';
+import { IgxGridComponent } from '../../grids/grid/grid.component';
 import { IgxExporterOptionsBase } from './exporter-options-base';
 import { IgxStringFilteringOperand } from '../../../public_api';
 import { GridIDNameJobTitleComponent } from '../../test-utils/grid-samples.spec';
