+++ projects/igniteui-angular/src/lib/test-utils/grid-base-components.spec.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit, ViewChild, AfterViewInit, ChangeDetectorRef } from '@angular/core';
 import { SampleTestData } from './sample-test-data.spec';
 import { ColumnDefinitions, GridTemplateStrings } from './template-strings.spec';
-import { IgxColumnHidingComponent, IgxColumnPinningComponent, IgxGridComponent } from '../grid';
+import { IgxColumnHidingComponent, IgxColumnPinningComponent } from '../grids';
+import { IgxGridComponent } from '../grids/grid/grid.component';
 
 @Component({
     template: `
